Make hanb read from stdin
