+++ rust/src/lib.rs
@@ -1,5 +1,7 @@
 const BOARD_SIZE: usize = 61;
 
+pub const DEFAULT_WIDTH: u8 = 12;
+
 /// Hanb cell. A cell has a fixed value and can resolve to a dynamic board.
 pub struct Cell {
     /// The value of the cell.
@@ -116,7 +118,7 @@ impl Navigator {
     }
 
     /// Go to a cell in the current board returning it
-    pub fn descend(&mut self, cell: usize) ->Result<&Board, String> {
+    pub fn descend(&mut self, cell: usize) -> Result<&Board, String> {
         self.resolve(cell)?;
         self.cell_stack.push(cell);
         Ok(self.current_board())
@@ -153,7 +155,8 @@ impl Navigator {
         if pos >= self.current_board().cells.len() {
             return Err(format!(
                 "Position must be less than {}. You entered {}",
-                self.current_board().cells.len(), pos
+                self.current_board().cells.len(),
+                pos
             ));
         }
 
@@ -164,4 +167,132 @@ impl Navigator {
         board.cells[pos].set_board_from_str(value)?;
         Ok(())
     }
+}
+
+pub fn print_board(board: &str, width: u8) {
+    if board.is_empty() || board.len() > 64 {
+        eprintln!(
+            "Input string must be 1 to 64 characters long. You entered with {}",
+            board.len()
+        );
+        return;
+    }
+    if width < 10 {
+        eprintln!("Width must be greater than {}", 9);
+        return;
+    }
+    let mut board = board.chars();
+    for c in 0..9 {
+        let mut n_cells = c + 5;
+        if c > 4 {
+            n_cells = 13 - c;
+        }
+        for r in 0..width {
+            if r < (width - n_cells) / 2 || r >= (width + n_cells) / 2 {
+                print!("▗");
+            } else {
+                if let Some(cell) = board.next() {
+                    print!("{}", cell);
+                } else {
+                    print!("▗");
+                }
+            }
+        }
+        println!();
+    }
+}
+
+pub fn eval(navigator: &mut Navigator, line: &str, stdout: bool) -> Result<(), String> {
+    // Process commands
+    let mut args = line.split_whitespace();
+    let command = args.next().unwrap();
+    let args = args.collect::<Vec<&str>>().join(" ");
+    match command {
+        "help" | "h" => {
+            println!("Commands:");
+            println!("  help, h: Print this help message");
+            println!("  print, p: Print the current board");
+            println!("  save, sv [filename]: Save the current board to a file");
+            println!("  load, ld [filename]: Load a board from a file");
+            println!("  down, d [cell_number]: Resolves into cell in the current board");
+            println!("  up, u: Ascends to containing board");
+            println!("  define, df [cell_number] [value]: Defines the inner board of a cell");
+        }
+        "print" | "p" => {
+            print_board(&navigator.root_board.to_string(), DEFAULT_WIDTH);
+        }
+        "save" | "sv" => return Err("Not implemented".to_string()),
+        "load" | "ld" => return Err("Not implemented".to_string()),
+        "down" | "d" => {
+            let cell = args
+                .parse::<usize>()
+                .map_err(|_| "Cell must be a number".to_string())?;
+            if cell >= navigator.root_board.cells.len() {
+                return Err(format!(
+                    "Cell must be between 0 and {}",
+                    navigator.root_board.cells.len() - 1
+                ));
+            }
+            match navigator.descend(cell) {
+                Ok(board) => {
+                    if stdout {
+                        println!(
+                            "Resolved cell '{}'. Board color '{}' You see:",
+                            cell, board.color
+                        );
+                        print_board(&board.to_string(), DEFAULT_WIDTH);
+                    }
+                }
+                Err(msg) => {
+                    return Err(msg);
+                }
+            }
+        }
+        "up" | "u" => match navigator.ascend() {
+            Some(board) => {
+                if stdout {
+                    println!("You ascend to the upper board. You see:");
+                    print_board(&board.to_string(), DEFAULT_WIDTH);
+                }
+            }
+            None => {
+                println!("{}", navigator.current_board().color);
+            }
+        },
+        "define" | "df" => {
+            let mut args = args.split_whitespace();
+            let cell = args
+                .next()
+                .unwrap()
+                .parse::<usize>()
+                .map_err(|_| "Cell must be a number".to_string())?;
+            if cell >= navigator.root_board.cells.len() {
+                return Err(format!(
+                    "Cell must be between 0 and {}",
+                    navigator.root_board.cells.len() - 1
+                ));
+            }
+            let value = args.collect::<Vec<&str>>().join(" ");
+            if value.len() > 64 {
+                return Err("Value must be 1 to 64 characters long".to_string());
+            }
+            match navigator.define(cell, value.as_str()) {
+                Ok(_) => {
+                    if stdout {
+                        print_board(&navigator.current_board().to_string(), DEFAULT_WIDTH);
+                    }
+                }
+                Err(msg) => {
+                    return Err(msg);
+                }
+            }
+        }
+        _ => {
+            return Err(format!(
+                "Unknown command: {}. Type help to see available commands.",
+                command
+            ));
+        }
+    }
+    Ok(())
 }
